Centralise current assemblies
